+++ src/deployment/RuleEngine.sol
@@ -4,7 +4,9 @@ pragma solidity ^0.8.20;
 
 /* ==== OpenZeppelin === */
 import {Context} from "OZ/utils/Context.sol";
+import {IAccessControl} from "OZ/access/IAccessControl.sol";
 import {AccessControl} from "OZ/access/AccessControl.sol";
+import {AccessControlEnumerable} from "OZ/access/extensions/AccessControlEnumerable.sol";
 import {IERC165} from "OZ/utils/introspection/ERC165.sol";
 /* ==== Modules === */
 import {ERC2771ModuleStandalone, ERC2771Context} from "../modules/ERC2771ModuleStandalone.sol";
@@ -14,7 +16,7 @@ import {RuleEngineBase} from "../RuleEngineBase.sol";
 /**
  * @title Implementation of a ruleEngine as defined by the CMTAT
  */
-contract RuleEngine is ERC2771ModuleStandalone, RuleEngineBase, AccessControl {
+contract RuleEngine is ERC2771ModuleStandalone, RuleEngineBase, AccessControlEnumerable {
     /**
      * @param admin Address of the contract (Access Control)
      * @param forwarderIrrevocable Address of the forwarder, required for the gasless support
@@ -36,7 +38,13 @@ contract RuleEngine is ERC2771ModuleStandalone, RuleEngineBase, AccessControl {
      * @notice Returns `true` if `account` has been granted `role`.
      * @dev The Default Admin has all roles
      */
-    function hasRole(bytes32 role, address account) public view virtual override(AccessControl) returns (bool) {
+    function hasRole(bytes32 role, address account)
+        public
+        view
+        virtual
+        override(AccessControl, IAccessControl)
+        returns (bool)
+    {
         if (AccessControl.hasRole(DEFAULT_ADMIN_ROLE, account)) {
             return true;
         } else {
@@ -45,9 +53,14 @@ contract RuleEngine is ERC2771ModuleStandalone, RuleEngineBase, AccessControl {
     }
 
     /* ============ ERC-165 ============ */
-    function supportsInterface(bytes4 interfaceId) public view virtual override(AccessControl, IERC165) returns (bool) {
-        return _supportsRuleEngineBaseInterface(interfaceId)
-            || AccessControl.supportsInterface(interfaceId);
+    function supportsInterface(bytes4 interfaceId)
+        public
+        view
+        virtual
+        override(AccessControlEnumerable, IERC165)
+        returns (bool)
+    {
+        return _supportsRuleEngineBaseInterface(interfaceId) || AccessControlEnumerable.supportsInterface(interfaceId);
     }
 
     /*//////////////////////////////////////////////////////////////