+++ test/RuleEngine/RuleEngineDeployment.t.sol
@@ -7,6 +7,7 @@ import "../HelperContract.sol";
 import {MinimalForwarderMock} from "CMTAT/mocks/MinimalForwarderMock.sol";
 import {IERC165} from "OZ/utils/introspection/IERC165.sol";
 import {IAccessControl} from "OZ/access/IAccessControl.sol";
+import {IAccessControlEnumerable} from "OZ/access/extensions/IAccessControlEnumerable.sol";
 import {ERC1404ExtendInterfaceId} from "CMTAT/library/ERC1404ExtendInterfaceId.sol";
 import {RuleEngineInterfaceId} from "CMTAT/library/RuleEngineInterfaceId.sol";
 import {ICompliance} from "src/mocks/ICompliance.sol";
@@ -59,6 +60,7 @@ contract RuleEngineTest is Test, HelperContract {
         // Act & Assert
         assertTrue(ruleEngineMock.supportsInterface(type(IERC165).interfaceId));
         assertTrue(ruleEngineMock.supportsInterface(type(IAccessControl).interfaceId));
+        assertTrue(ruleEngineMock.supportsInterface(type(IAccessControlEnumerable).interfaceId));
         assertTrue(ruleEngineMock.supportsInterface(RuleEngineInterfaceId.RULE_ENGINE_INTERFACE_ID));
         assertTrue(ruleEngineMock.supportsInterface(ERC1404ExtendInterfaceId.ERC1404EXTEND_INTERFACE_ID));
         assertTrue(ruleEngineMock.supportsInterface(type(ICompliance).interfaceId));